// scripts/Pix-Scripting/src/types.ts
export interface PixivImageUrls {
  square_medium?: string
  medium?: string
  large?: string
  original?: string
}

export interface PixivTag {
  name: string
  translated_name?: string | null
}

export interface PixivUser {
  id: number
  name: string
  account: string
  profile_image_urls?: { medium?: string }
  comment?: string
  is_followed?: boolean
  is_mypixiv?: boolean
}

export interface PixivSeries {
  id: number
  title?: string | null
}

export interface PixivWatchlistSeries {
  id: number
  title: string
  url?: string | null
  mask_text?: string | null
  published_content_count: number
  last_published_content_datetime?: string | null
  latest_content_id?: number | null
  user?: PixivUser | null
}

export interface PixivMetaPage {
  image_urls: PixivImageUrls
}

export interface PixivMetaSinglePage {
  original_image_url?: string
}

export interface PixivIllustration {
  id: number
  title: string
  type: "illust" | "manga" | "ugoira"
  image_urls: PixivImageUrls
  caption: string
  user: PixivUser
  tags: PixivTag[]
  create_date: string
  page_count: number
  width: number
  height: number
  x_restrict: number
  series?: PixivSeries | null
  episode_number?: number
  meta_single_page?: PixivMetaSinglePage
  meta_pages: PixivMetaPage[]
  total_view: number
  total_bookmarks: number
  is_bookmarked: boolean
  is_muted: boolean
  illust_ai_type: number
  total_comments: number
  comment_access_control: number
}
// ...
export interface PixivUserPreview {
  user: PixivUser
  illusts: PixivIllustration[]
  novels?: PixivNovel[]
  is_muted: boolean
}
// ...
export interface PixivNovel {
  id: number
  title: string
  caption?: string
  user: PixivUser
  tags: PixivTag[]
  create_date: string
  page_count: number
  x_restrict: number
  total_view: number
  total_bookmarks: number
  is_bookmarked: boolean
  is_muted: boolean
  novel_ai_type?: number
  total_comments: number
  text_length?: number
  visible: boolean
  series?: PixivSeries | null
  episode_number?: number
  novel_marker?: {
    page: number
  } | null
  image_urls?: {
    square_medium?: string
    medium?: string
    large?: string
  } | null
  cover?: {
    urls?: { "240mw"?: string; "480mw"?: string; "1200x1200"?: string }
  } | null
}